tests/wpt/web-platform-tests/element-timing/observe-text.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Element Timing: observe text</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
  async_test((t) => {
    if (!window.PerformanceElementTiming) {
      assert_unreached("PerformanceElementTiming is not implemented");
    }
    let paragraph;
    let beforeRender;
    const observer = new PerformanceObserver(
      t.step_func_done((entryList) => {
        assert_equals(entryList.getEntries().length, 1);
        checkTextElement(entryList.getEntries()[0], 'my_text', 'text_id', beforeRender, paragraph);
      })
    );
    observer.observe({entryTypes: ['element']});
    // We add the iframe during onload to be sure that the observer is registered
    // in time for it to observe the element timing.
    window.onload = () => {
      paragraph = document.createElement('p');
      paragraph.innerHTML = 'This is text I care about';
      paragraph.setAttribute('elementtiming', 'my_text');
      paragraph.setAttribute('id', 'text_id');
      document.body.appendChild(paragraph);
      beforeRender = performance.now();
    };
  }, 'Paragraph with elementtiming attribute is observed.');
</script>

</body>
